# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scholars Essentials API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type ProfileUserType =
  (typeof ProfileUserType)[keyof typeof ProfileUserType];

export const ProfileUserType = {
  student: "student",
  owner: "owner",
} as const;

export interface Profile {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  userType: ProfileUserType;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
  /** @nullable */
  city: string | null;
}

export type UpdateProfileBodyUserType =
  (typeof UpdateProfileBodyUserType)[keyof typeof UpdateProfileBodyUserType];

export const UpdateProfileBodyUserType = {
  student: "student",
  owner: "owner",
} as const;

export interface UpdateProfileBody {
  userType?: UpdateProfileBodyUserType;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  city?: string | null;
}

export type ItemCondition = (typeof ItemCondition)[keyof typeof ItemCondition];

export const ItemCondition = {
  new: "new",
  like_new: "like_new",
  good: "good",
  fair: "fair",
  used: "used",
} as const;

export interface Item {
  id: number;
  sellerId: string;
  /** @nullable */
  sellerName: string | null;
  /** @nullable */
  sellerEmail?: string | null;
  title: string;
  description: string;
  price: number;
  condition: ItemCondition;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export type CreateItemBodyCondition =
  (typeof CreateItemBodyCondition)[keyof typeof CreateItemBodyCondition];

export const CreateItemBodyCondition = {
  new: "new",
  like_new: "like_new",
  good: "good",
  fair: "fair",
  used: "used",
} as const;

export interface CreateItemBody {
  /** @minLength 1 */
  title: string;
  description: string;
  /** @minimum 0 */
  price: number;
  condition: CreateItemBodyCondition;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export type UpdateItemBodyCondition =
  (typeof UpdateItemBodyCondition)[keyof typeof UpdateItemBodyCondition];

export const UpdateItemBodyCondition = {
  new: "new",
  like_new: "like_new",
  good: "good",
  fair: "fair",
  used: "used",
} as const;

export interface UpdateItemBody {
  /** @minLength 1 */
  title?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  condition?: UpdateItemBodyCondition;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

/**
 * @nullable
 */
export type AccommodationGenderPreference =
  | (typeof AccommodationGenderPreference)[keyof typeof AccommodationGenderPreference]
  | null;

export const AccommodationGenderPreference = {
  any: "any",
  male: "male",
  female: "female",
} as const;

export interface Accommodation {
  id: number;
  ownerId: string;
  /** @nullable */
  ownerName: string | null;
  /** @nullable */
  ownerEmail?: string | null;
  name: string;
  address: string;
  /** @nullable */
  description?: string | null;
  monthlyRent: number;
  latitude: number;
  longitude: number;
  /** @nullable */
  genderPreference?: AccommodationGenderPreference;
  amenities?: string[];
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

/**
 * @nullable
 */
export type CreateAccommodationBodyGenderPreference =
  | (typeof CreateAccommodationBodyGenderPreference)[keyof typeof CreateAccommodationBodyGenderPreference]
  | null;

export const CreateAccommodationBodyGenderPreference = {
  any: "any",
  male: "male",
  female: "female",
} as const;

export interface CreateAccommodationBody {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  address: string;
  /** @nullable */
  description?: string | null;
  /** @minimum 0 */
  monthlyRent: number;
  latitude: number;
  longitude: number;
  /** @nullable */
  genderPreference?: CreateAccommodationBodyGenderPreference;
  amenities?: string[];
  /** @nullable */
  imageUrl?: string | null;
}

/**
 * @nullable
 */
export type UpdateAccommodationBodyGenderPreference =
  | (typeof UpdateAccommodationBodyGenderPreference)[keyof typeof UpdateAccommodationBodyGenderPreference]
  | null;

export const UpdateAccommodationBodyGenderPreference = {
  any: "any",
  male: "male",
  female: "female",
} as const;

export interface UpdateAccommodationBody {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  address?: string;
  /** @nullable */
  description?: string | null;
  /** @minimum 0 */
  monthlyRent?: number;
  latitude?: number;
  longitude?: number;
  /** @nullable */
  genderPreference?: UpdateAccommodationBodyGenderPreference;
  amenities?: string[];
  /** @nullable */
  imageUrl?: string | null;
}

export interface MapLocation {
  id: number;
  name: string;
  address: string;
  latitude: number;
  longitude: number;
  monthlyRent: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface DashboardSummary {
  totalItems: number;
  totalAccommodations: number;
  totalUsers: number;
  averageItemPrice: number;
  averageMonthlyRent: number;
}

export type ActivityEntryKind =
  (typeof ActivityEntryKind)[keyof typeof ActivityEntryKind];

export const ActivityEntryKind = {
  item: "item",
  accommodation: "accommodation",
} as const;

export interface ActivityEntry {
  id: string;
  kind: ActivityEntryKind;
  title: string;
  subtitle: string;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface ConditionCount {
  condition: string;
  count: number;
}

export interface RentStats {
  min: number;
  max: number;
  avg: number;
  count: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListItemsParams = {
  q?: string;
  condition?: ListItemsCondition;
};

export type ListItemsCondition =
  (typeof ListItemsCondition)[keyof typeof ListItemsCondition];

export const ListItemsCondition = {
  new: "new",
  like_new: "like_new",
  good: "good",
  fair: "fair",
  used: "used",
} as const;

export type ListAccommodationsParams = {
  q?: string;
  maxRent?: number;
};