# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Scholars Essentials API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      tags: [profile]
      operationId: getProfile
      summary: Get my full profile (user_type, location, etc.)
      responses:
        '200':
          description: Profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: Unauthorized
    put:
      tags: [profile]
      operationId: updateProfile
      summary: Update my profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileBody'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: Unauthorized
  /items:
    get:
      tags: [items]
      operationId: listItems
      summary: List items (books / essentials) for sale
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: condition
          in: query
          required: false
          schema:
            type: string
            enum: [new, like_new, good, fair, used]
      responses:
        '200':
          description: List of items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
    post:
      tags: [items]
      operationId: createItem
      summary: Create an item listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemBody'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '401':
          description: Unauthorized
  /items/mine:
    get:
      tags: [items]
      operationId: listMyItems
      summary: List the current user's item listings
      responses:
        '200':
          description: Items owned by current user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '401':
          description: Unauthorized
  /items/{id}:
    get:
      tags: [items]
      operationId: getItem
      summary: Get a single item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Not found.
    put:
      tags: [items]
      operationId: updateItem
      summary: Update one of your item listings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemBody'
      responses:
        '200':
          description: Updated item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
    delete:
      tags: [items]
      operationId: deleteItem
      summary: Delete one of your item listings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /accommodations:
    get:
      tags: [accommodations]
      operationId: listAccommodations
      summary: List PG accommodations
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: maxRent
          in: query
          required: false
          schema:
            type: number
      responses:
        '200':
          description: List of accommodations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Accommodation'
    post:
      tags: [accommodations]
      operationId: createAccommodation
      summary: Create a PG listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccommodationBody'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accommodation'
        '401':
          description: Unauthorized
  /accommodations/mine:
    get:
      tags: [accommodations]
      operationId: listMyAccommodations
      summary: List the current user's PG listings
      responses:
        '200':
          description: Accommodations owned by current user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Accommodation'
        '401':
          description: Unauthorized
  /accommodations/{id}:
    get:
      tags: [accommodations]
      operationId: getAccommodation
      summary: Get a single PG
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Accommodation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accommodation'
        '404':
          description: Not found.
    put:
      tags: [accommodations]
      operationId: updateAccommodation
      summary: Update a PG listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccommodationBody'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accommodation'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
    delete:
      tags: [accommodations]
      operationId: deleteAccommodation
      summary: Delete a PG listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
...
components:
...
  schemas:
...
    Profile:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, userType, latitude, longitude, city]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        userType:
          type: string
          enum: [student, owner]
        latitude:
          type: [number, 'null']
        longitude:
          type: [number, 'null']
        city:
          type: [string, 'null']
    UpdateProfileBody:
      type: object
      properties:
        userType:
          type: string
          enum: [student, owner]
        latitude:
          type: [number, 'null']
        longitude:
          type: [number, 'null']
        city:
          type: [string, 'null']
    Item:
      type: object
      required: [id, sellerId, sellerName, title, description, price, condition, createdAt]
      properties:
        id:
          type: integer
        sellerId:
          type: string
        sellerName:
          type: [string, 'null']
        sellerEmail:
          type: [string, 'null']
        title:
          type: string
        description:
          type: string
        price:
          type: number
        condition:
          type: string
          enum: [new, like_new, good, fair, used]
        category:
          type: [string, 'null']
        imageUrl:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    CreateItemBody:
      type: object
      required: [title, description, price, condition]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          minimum: 0
        condition:
          type: string
          enum: [new, like_new, good, fair, used]
        category:
          type: [string, 'null']
        imageUrl:
          type: [string, 'null']
    UpdateItemBody:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          minimum: 0
        condition:
          type: string
          enum: [new, like_new, good, fair, used]
        category:
          type: [string, 'null']
        imageUrl:
          type: [string, 'null']
    Accommodation:
      type: object
      required: [id, ownerId, ownerName, name, address, monthlyRent, latitude, longitude, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: string
        ownerName:
          type: [string, 'null']
        ownerEmail:
          type: [string, 'null']
        name:
          type: string
        address:
          type: string
        description:
          type: [string, 'null']
        monthlyRent:
          type: number
        latitude:
          type: number
        longitude:
          type: number
        genderPreference:
          type: [string, 'null']
          enum: [any, male, female, null]
        amenities:
          type: array
          items:
            type: string
        imageUrl:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    CreateAccommodationBody:
      type: object
      required: [name, address, monthlyRent, latitude, longitude]
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        description:
          type: [string, 'null']
        monthlyRent:
          type: number
          minimum: 0
        latitude:
          type: number
        longitude:
          type: number
        genderPreference:
          type: [string, 'null']
          enum: [any, male, female, null]
        amenities:
          type: array
          items:
            type: string
        imageUrl:
          type: [string, 'null']
    UpdateAccommodationBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        description:
          type: [string, 'null']
        monthlyRent:
          type: number
          minimum: 0
        latitude:
          type: number
        longitude:
          type: number
        genderPreference:
          type: [string, 'null']
          enum: [any, male, female, null]
        amenities:
          type: array
          items:
            type: string
        imageUrl:
          type: [string, 'null']
...